requirements.txt
Flask==1.0.2

PyMySQL==0.9.3

pyzmq==17.1.2

mysqlclient==1.4.2.post1
